+++ src/main/java/org/zhadok/poe/controller/util/Point.java
@@ -2,10 +2,10 @@ package org.zhadok.poe.controller.util;
 
 public class Point {
 	
-	public final int x; 
-	public final int y; 
+	public final double x; 
+	public final double y; 
 	
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		this.x = x; 
 		this.y = y; 
 	}
@@ -20,6 +20,17 @@ public class Point {
 				Math.pow(target.x-this.x, 2) + 
 				Math.pow(target.y-this.y, 2)
 				);
+	}
+
+	/**
+	 * Assuming the result is a point on the unit circle
+	 * https://www.xarg.org/2017/07/how-to-map-a-square-to-a-circle/
+	 * @return
+	 */
+	public Point mapFromCircleToSquare() {
+		double xOnCircle = x * Math.sqrt(1 - y * y / 2); 
+		double yOnCircle = y * Math.sqrt(1 - x * x / 2);
+		return new Point(xOnCircle, yOnCircle);
 	}
 	
 }