+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+- 0.0.4: When listening for new controller input: 
+-- Fixed a bug with the cmd file for starting the jar
+-- Character movement: Map point on square to point on unit circle for joysticks
+
 - 0.0.3: Automated deployment via GitHub Actions
 - 0.0.2: Update deployment and versioning
 - 0.0.1: Initial release